requirements.txt
fastapi==0.100.0
uvicorn==0.15.0
pydantic==2.0.3
celery==5.4.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
redis==5.0.7
requests==2.32.3
SQLAlchemy==2.0.31
python-dotenv==1.0.1